+++ extensions/badges/src/createSocialBadge.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { Action, ActionPanel, Cache, Detail, Icon, LaunchProps, getPreferenceValues } from "@raycast/api";
+import { Action, ActionPanel, Detail, Icon, LaunchProps, getPreferenceValues } from "@raycast/api";
 import { useCachedState } from "@raycast/utils";
 import { omitBy } from "lodash";
 import { Documentation } from "./components/actions.js";
@@ -8,9 +8,6 @@ import { Input } from "./components/input.js";
 import { colorsForBackground } from "./vendor/badge-maker-color.js";
 import { Badge, LaunchFromSimpleIconsContext, LaunchFromColorPickerContext } from "./types.js";
 import { codeBlock, encodeBadgeContentParameters } from "./utils.js";
-
-const cache = new Cache();
-cache.set("launchCommandName", JSON.stringify("createSocialBadge"));
 
 const defaultBadge: Badge = {
   $icon: { title: "Raycast", slug: "raycast", hex: "FF6363", source: "" },
@@ -51,7 +48,6 @@ export default function Command({
     }
 
     if (launchContext?.launchFromExtensionName === "color-picker" && launchContext?.hex && launchContext?.field) {
-      console.log(launchContext);
       setBadge({ ...badge, [launchContext.field]: launchContext.hex.slice(1) });
     }
   }, []);